+++ solution/24/100Years.html
@@ -1,0 +1,23 @@
+<div class="card" brd-20>
+
+    <p>
+        <h1 id="the-next-100-years">The Next 100 Years</h1>
+        <p>A Forecast for the 21st Century by George Friedman</p>
+        <h2 id="amazon.com">Amazon.com</h2>
+        <p><a href="https://www.amazon.com/Next-100-Years-Forecast-Century-ebook/dp/B001NLL946"><img src="https://images-na.ssl-images-amazon.com/images/I/51g3kMcyksL.jpg" /></a></p>
+        <dl>
+            <dt><a href="https://www.amazon.com/Next-100-Years-Forecast-Century-ebook/dp/B001NLL946">The Next 100 Years</a></dt>
+            <dd>A Forecast for the 21st Century by George Friedman
+            </dd>
+        </dl>
+        <hr />
+        <h2 id="overview">Overview</h2>
+        <p>George Friedman runs a private intelligence service. His analysis of the key issues of geopolitics leads him to identify key problems that must be solved by nations of the world. There are a limited number of options that are practical for each nations leaders. This sets certain events in motion that will dictate future behavior.</p>
+        <p>The predictions here are based on thorough analysis of the data for economics, demographics, and geography for the people of the world. He lays out the history of the last 500 years and shows how many of the major turning points could have been predicted by understanding the geopolitical issues involved.</p>
+        <p>This is one of the most fascinating books I have ever read. It helped me understand world history by looking at how events unfold. The point is not to predict the future but to better understand it.</p>
+        <h2 id="excerpt">Excerpt</h2>
+        <p>When you look at a map of China closely, you see that it is really a very isolated country physically. China has not been a major naval power for centuries, and building a navy requires a long time. There is a deeper reason for not worrying about China. China is inherently unstable. Whenever it opens its borders to the outside world, the coastal region becomes prosperous, but the vast majority of Chinese in the interior remain impoverished. This leads to tension, conflict, and instability. It also leads to economic decisions made for political reasons, resulting in inefficiency and corruption.</p>
+
+        <p>Read more <a href="http://mybookonline.org/booknotes/Next100Years">http://mybookonline.org/booknotes/Next100Years</a></p>
+
+</div>
