+++ solution/24/page.html
@@ -12,7 +12,7 @@
 
         <header>
             <img src="Bear.png" alt="Bear Logo"/>
-            <h1>{{ site_title }} $ . '</h1>
+            <h1>{{ site_title }}</h1>
             <h2>{{ page_title }}</h2>
         </header>
         <main>
