+++ templates/views.php
@@ -93,6 +93,14 @@
 
 
     /*
+        render_page -- Create one HTML page from a template.
+    */
+    function render_page($settings) {
+        return render_template("page.html", $settings);
+    }
+
+
+    /*
         render_simple_page -- Create the HTML page.
     */
     function render_simple_page($title, $text) {
@@ -117,7 +125,7 @@
     function transform_text ($text, $settings) {
         foreach ($settings as $key => $value) {
             // echo "$key => $value"; 
-            $text = str_replace($key,  $value,  $text);
+            $text = str_replace("{{ $key }}",  $value,  $text);
         }
         return $text;
     }
